requirements.txt
appdirs==1.4.3
asn1crypto==0.24.0
attrs==18.2.0
cached-property==1.5.1
certifi==2018.11.29
cffi==1.12.1
chardet==3.0.4
cryptography==2.5
defusedxml==0.5.0
Django==1.11.3
django-recaptcha==1.0.5
djangorestframework==3.6.3
djangorestframework-xml==1.4.0
eight==0.4.2
future==0.16.0
gevent==1.2.2
google-api-python-client==1.6.2
greenlet==0.4.15
gunicorn==19.9.0
httplib2==0.12.1
idna==2.8
ISO8583==1.3.4
isodate==0.6.0
lxml==4.3.1
M2Crypto==0.31.0
mysqlclient==1.3.10
oauth2client==4.1.3
pkgconfig==1.4.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
pycryptodome==3.4.6
pyOpenSSL==18.0.0
python-memcached==1.58
pytz==2018.9
requests==2.21.0
requests-toolbelt==0.9.1
rsa==4.0
signxml==2.6.0
six==1.12.0
suds-py3==1.3.3.0
uritemplate==3.0.0
urllib3==1.24.1
XlsxWriter==1.1.2
xlwt==1.3.0
xmlsec==1.3.3
zeep==3.1.0
redis==3.2.1
pandas==0.24.2
oauthlib==3.0.1
dicttoxml==1.7.4
xmltodict=0.12.0
mock==3.0.4
coverage==4.5.3
